requirements.txt
Django==2
djangorestframework==3.8.2
django-filter==1.1.0
python-decouple=3.1